+++ R/translate-sql-odbc.R
@@ -1,4 +1,3 @@
-#' @include translate-sql-paste.R
 #' @export
 #' @rdname sql_variant
 #' @format NULL
@@ -6,26 +5,7 @@ base_odbc_scalar <- sql_translator(.parent = base_scalar,
   as.numeric    = sql_cast("DOUBLE"),
   as.double     = sql_cast("DOUBLE"),
   as.integer    = sql_cast("INT"),
-  # bit64::as.integer64 can translate to BIGINT data type:
-  # MS SQL - https://docs.microsoft.com/en-us/sql/t-sql/data-types/int-bigint-smallint-and-tinyint-transact-sql
-  # Hive - https://cwiki.apache.org/confluence/display/Hive/LanguageManual+Types#LanguageManualTypes-IntegralTypes(TINYINT,SMALLINT,INT/INTEGER,BIGINT)
-  # Postgres - https://www.postgresql.org/docs/8.4/static/datatype-numeric.html
-  # Impala - https://impala.apache.org/docs/build/html/topics/impala_bigint.html
-  as.integer64  = sql_cast("BIGINT"),
-  as.logical    = sql_cast("BOOLEAN"),
-  as.character  = sql_cast("STRING"),
-  as.Date       = sql_cast("DATE"),
-  paste0        = sql_prefix("CONCAT"),
-                  # cosh, sinh, coth and tanh calculations are based on this article
-                  # https://en.wikipedia.org/wiki/Hyperbolic_function
-  round         = function(x, digits = 0L){
-                    sql_expr(ROUND(!!x, !!as.integer(digits)))
-                  },
-  cosh          = function(x) sql_expr((!!sql_exp(1, x) + !!sql_exp(-1, x)) / 2L),
-  sinh          = function(x) sql_expr((!!sql_exp(1, x) - !!sql_exp(-1, x)) / 2L),
-  tanh          = function(x) sql_expr((!!sql_exp(2, x) - 1L) / (!!sql_exp(2, x) + 1L)),
-  coth          = function(x) sql_expr((!!sql_exp(2, x) + 1L) / (!!sql_exp(2, x) - 1L)),
-  paste         = sql_paste(" ")
+  as.character  = sql_cast("STRING")
 )
 
 sql_exp <- function(a, x) {